+++ src/main/java/net/sf/jabref/external/push/PushToApplication.java
@@ -13,7 +13,7 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
-package net.sf.jabref.external;
+package net.sf.jabref.external.push;
 
 import net.sf.jabref.gui.BasePanel;
 import net.sf.jabref.model.database.BibtexDatabase;
