+++ src/main/java/net/sf/jabref/external/push/PushToLatexEditor.java
@@ -13,13 +13,14 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
-package net.sf.jabref.external;
+package net.sf.jabref.external.push;
 
 import java.io.IOException;
 
 import javax.swing.*;
 
 import net.sf.jabref.*;
+
 import com.jgoodies.forms.builder.FormBuilder;
 import com.jgoodies.forms.layout.FormLayout;
 import net.sf.jabref.gui.BasePanel;
@@ -30,25 +31,25 @@ import net.sf.jabref.model.database.BibtexDatabase;
 import net.sf.jabref.model.entry.BibtexEntry;
 
 /**
- * Class for pushing entries into TexMaker.
+ * Class for pushing entries into LatexEditor.
  */
-public class PushToTexmaker implements PushToApplication {
+public class PushToLatexEditor implements PushToApplication {
 
     private boolean couldNotCall;
     private boolean notDefined;
     private JPanel settings;
-    private final JTextField texmakerPath = new JTextField(30);
+    private final JTextField ledPath = new JTextField(30);
     private final JTextField citeCommand = new JTextField(30);
 
 
     @Override
     public String getName() {
-        return Localization.lang("Insert selected citations into %0", getApplicationName());
+        return Localization.lang("Insert selected citations into %0" ,getApplicationName());
     }
 
     @Override
     public String getApplicationName() {
-        return "Texmaker";
+        return "LatexEditor";
     }
 
     @Override
@@ -58,7 +59,7 @@ public class PushToTexmaker implements PushToApplication {
 
     @Override
     public Icon getIcon() {
-        return IconTheme.getImage("texmaker");
+        return IconTheme.getImage("edit");
     }
 
     @Override
@@ -72,15 +73,15 @@ public class PushToTexmaker implements PushToApplication {
         couldNotCall = false;
         notDefined = false;
 
-        String texMaker = Globals.prefs.get(JabRefPreferences.TEXMAKER_PATH);
+        String led = Globals.prefs.get(JabRefPreferences.LATEX_EDITOR_PATH);
 
-        if ((texMaker == null) || texMaker.trim().isEmpty()) {
+        if (led == null || led.trim().isEmpty()) {
             notDefined = true;
             return;
         }
 
         try {
-            Runtime.getRuntime().exec(texMaker + " " + "-insert " + Globals.prefs.get(JabRefPreferences.CITE_COMMAND_TEXMAKER) + "{" + keyString + "}");
+            Runtime.getRuntime().exec(led + " " + "-i " + Globals.prefs.get(JabRefPreferences.CITE_COMMAND_LED) + "{" + keyString + "}");
 
         }
 
@@ -94,14 +95,17 @@ public class PushToTexmaker implements PushToApplication {
     public void operationCompleted(BasePanel panel) {
         if (notDefined) {
             // @formatter:off
-            panel.output(Localization.lang("Error") + ": " 
-                    + Localization.lang("Path to %0 not defined", getApplicationName()) + ".");
+            panel.output(Localization.lang("Error") + ": " +
+                    Localization.lang("Path to %0 not defined", getApplicationName()) + ".");
+            // @formatter:on
         } else if (couldNotCall) {
-            panel.output(Localization.lang("Error") + ": "
-                    + Localization.lang("Could not call executable") + " '" + Globals.prefs.get(JabRefPreferences.TEXMAKER_PATH) + "'.");
+            // @formatter:off
+            panel.output(Localization.lang("Error") + ": " + 
+                    Localization.lang("Could not call executable") + " '"
+                    + Globals.prefs.get(JabRefPreferences.LATEX_EDITOR_PATH) + "'.");
             // @formatter:on
         } else {
-            panel.output(Localization.lang("Pushed citations to %0", getApplicationName()));
+            Localization.lang("Pushed citations to %0", "LatexEditor");
         }
     }
 
@@ -115,28 +119,28 @@ public class PushToTexmaker implements PushToApplication {
         if (settings == null) {
             initSettingsPanel();
         }
-        texmakerPath.setText(Globals.prefs.get(JabRefPreferences.TEXMAKER_PATH));
-        citeCommand.setText(Globals.prefs.get(JabRefPreferences.CITE_COMMAND_TEXMAKER));
+        ledPath.setText(Globals.prefs.get(JabRefPreferences.LATEX_EDITOR_PATH));
+        citeCommand.setText(Globals.prefs.get(JabRefPreferences.CITE_COMMAND_LED));
         return settings;
     }
 
     private void initSettingsPanel() {
         FormBuilder builder = FormBuilder.create();
         builder.layout(new FormLayout("left:pref, 4dlu, fill:pref:grow, 4dlu, fill:pref", "p, 2dlu, p"));
-        builder.add(Localization.lang("Path to %0", getApplicationName()) + ":").xy(1, 1);
-        builder.add(texmakerPath).xy(3, 1);
-        BrowseAction action = BrowseAction.buildForFile(texmakerPath);
+        builder.add(Localization.lang("Path to LatexEditor (LEd.exe)") + ":").xy(1, 1); // Note the LEd.exe part
+        builder.add(ledPath).xy(3,1);
+        BrowseAction action = BrowseAction.buildForFile(ledPath);
         JButton browse = new JButton(Localization.lang("Browse"));
         browse.addActionListener(action);
-        builder.add(browse).xy(5, 1);
+        builder.add(browse).xy(5,1);
         builder.add(Localization.lang("Cite command") + ":").xy(1, 3);
-        builder.add(citeCommand).xy(3, 3);
+        builder.add(citeCommand).xy(3,3);
         settings = builder.build();
     }
 
     @Override
     public void storeSettings() {
-        Globals.prefs.put(JabRefPreferences.TEXMAKER_PATH, texmakerPath.getText());
-        Globals.prefs.put(JabRefPreferences.CITE_COMMAND_TEXMAKER, citeCommand.getText());
+        Globals.prefs.put(JabRefPreferences.LATEX_EDITOR_PATH, ledPath.getText());
+        Globals.prefs.put(JabRefPreferences.CITE_COMMAND_LED, citeCommand.getText());
     }
 }
