+++ src/main/java/net/sf/jabref/external/push/PushToWinEdt.java
@@ -13,13 +13,14 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
-package net.sf.jabref.external;
+package net.sf.jabref.external.push;
 
 import java.io.IOException;
 
 import javax.swing.*;
 
 import net.sf.jabref.*;
+
 import com.jgoodies.forms.builder.FormBuilder;
 import com.jgoodies.forms.layout.FormLayout;
 import net.sf.jabref.gui.BasePanel;
