+++ src/main/java/net/sf/jabref/gui/JabRefFrame.java
@@ -73,7 +73,7 @@ import net.sf.jabref.exporter.ExportFormats;
 import net.sf.jabref.exporter.SaveAllAction;
 import net.sf.jabref.exporter.SaveDatabaseAction;
 import net.sf.jabref.external.ExternalFileTypeEditor;
-import net.sf.jabref.external.PushToApplicationButton;
+import net.sf.jabref.external.push.PushToApplicationButton;
 import net.sf.jabref.groups.EntryTableTransferHandler;
 import net.sf.jabref.groups.GroupSelector;
 import net.sf.jabref.gui.menus.help.ForkMeOnGitHubAction;
