+++ src/main/java/net/sf/jabref/openoffice/OpenOfficePanel.java
@@ -21,7 +21,7 @@ import com.jgoodies.forms.layout.FormLayout;
 import net.sf.jabref.*;
 import net.sf.jabref.exporter.layout.Layout;
 import net.sf.jabref.exporter.layout.LayoutHelper;
-import net.sf.jabref.external.PushToApplication;
+import net.sf.jabref.external.push.PushToApplication;
 import net.sf.jabref.gui.*;
 import net.sf.jabref.gui.worker.AbstractWorker;
 import net.sf.jabref.gui.actions.BrowseAction;
